p-values formatter

Formatter for p-values, adding a symbol "<" for small p-values.

pvalue(x, accuracy = 0.001, decimal.mark = ".", prefix = NULL, add_p = FALSE)

label_pvalue(
  accuracy = 0.001,
  decimal.mark = ".",
  prefix = NULL,
  add_p = FALSE
)

signif_stars(x, three = 0.001, two = 0.01, one = 0.05, point = 0.1)

Arguments

x
A numeric vector to format.

accuracy
A number to round to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, the default, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values.

Applied to rescaled data.

decimal.mark
The character to be used to indicate the numeric decimal point.

prefix
A character vector of length 3 giving the prefixes to put in front of numbers. The default values are c("<", "", ">") if add_p is TRUE and c("p<", "p=", "p>") if FALSE.

add_p
Add "p=" before the value?

three
threshold below which to display three stars

two
threshold below which to display two stars

one
threshold below which to display one star

point
threshold below which to display one point (NULL to desactivate)

Value

label_pvalue returns a function with single parameter x, a numeric vector, that returns a character vector.

Examples

p <- c(.50, 0.12, .045, .011, .009, .00002, NA)
pvalue(p)

#> [1] "0.500"  "0.120"  "0.045"  "0.011"  "0.009"  "<0.001" NA      

pvalue(p, accuracy = .01)

#> [1] "0.50"  "0.12"  "0.04"  "0.01"  "<0.01" "<0.01" NA     

pvalue(p, add_p = TRUE)

#> [1] "p=0.500" "p=0.120" "p=0.045" "p=0.011" "p=0.009" "p<0.001" NA       

custom_function <- label_pvalue(accuracy = .1, decimal.mark = ",")
custom_function(p)

#> [1] "0,5"  "0,1"  "<0,1" "<0,1" "<0,1" "<0,1" NA    

signif_stars(p)

#> [1] ""    ""    "*"   "*"   "**"  "***" ""   

signif_stars(p, one = .15, point = NULL)

#> [1] ""    "*"   "*"   "*"   "**"  "***" ""