Tidy data with variable names and factor levels

For a more recent and powerful function, use broom.helpers::tidy_plus_plus()

tidy_detailed(x, ...)

Arguments

x
model to tidy

...
extra arguments passed to tidy function from broom package

Source

This function has been adapted from Benjamin Nutter's tidy_levels_labels function implemented within the pixiedust package (https://github.com/nutterb/pixiedust/blob/master/R/tidy_levels_labels.R).

Value

The data.frame produced by tidy with these additional columns:

variable

variable name

variable_label

label of the variable (if available), variable name otherwise

level

for factors, the corresponding level

level_detail

level with indication of the reference level, eg. B vs. A

label

equal to level_detail if it's a factor, to variable_label otherwise

Note

This function requires the broom and dplyr packages.

See also

Examples

reg <- lm(Sepal.Length ~ Sepal.Width + Petal.Length * Species, data = iris)
tidy_detailed(reg)

#> Warning: `tidy_detailed()` is deprecated as of JLutils 1.22.0.
#> Please use `broom.helpers::tidy_plus_plus()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Warning: `mutate_()` is deprecated as of dplyr 0.7.0.
#> Please use `mutate()` instead.
#> See vignette('programming') for more help
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> # A tibble: 7 x 10
#>   term  estimate std.error statistic  p.value variable variable_label level
#>   <chr>    <dbl>     <dbl>     <dbl>    <dbl> <chr>    <chr>          <chr>
#> 1 (Int…    2.93     0.436       6.70 4.34e-10 NA       NA              NA  
#> 2 Sepa…    0.450    0.0806      5.58 1.15e- 7 Sepal.W… Sepal.Width    ""   
#> 3 Peta…    0.368    0.254       1.45 1.49e- 1 Petal.L… Petal.Length   ""   
#> 4 Spec…   -1.05     0.561      -1.87 6.39e- 2 Species  Species        "ver…
#> 5 Spec…   -2.62     0.584      -4.48 1.51e- 5 Species  Species        "vir…
#> 6 Peta…    0.292    0.268       1.09 2.78e- 1 Petal.L… Petal.Length:… "Pet…
#> 7 Peta…    0.523    0.264       1.98 4.98e- 2 Petal.L… Petal.Length:… "Pet…
#> # … with 2 more variables: level_detail <chr>, label <chr>

d <- as.data.frame(Titanic)
reg2 <- glm(Survived ~ Sex + Age + Class, family = binomial, data = d, weights = d$Freq)
tidy_detailed(reg2, conf.int = TRUE, exponentiate = TRUE)

#> # A tibble: 6 x 12
#>   term  estimate std.error statistic  p.value conf.low conf.high variable
#>   <chr>    <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl> <chr>   
#> 1 (Int…    1.98      0.273      2.51 1.21e- 2    1.16      3.39  NA      
#> 2 SexF…   11.2       0.140     17.2  1.43e-66    8.57     14.9   Sex     
#> 3 AgeA…    0.346     0.244     -4.35 1.36e- 5    0.214     0.558 Age     
#> 4 Clas…    0.361     0.196     -5.19 2.05e- 7    0.245     0.529 Class   
#> 5 Clas…    0.169     0.172    -10.4  3.69e-25    0.120     0.236 Class   
#> 6 Clas…    0.424     0.157     -5.45 5.00e- 8    0.312     0.577 Class   
#> # … with 4 more variables: variable_label <chr>, level <chr>,
#> #   level_detail <chr>, label <chr>