Tidy data with variable names and factor levels

For a more recent and powerful function, use broom.helpers::tidy_plus_plus()

tidy_detailed(x, ...)

Arguments

x
model to tidy

...
extra arguments passed to tidy function from broom package

Source

This function has been adapted from Benjamin Nutter's tidy_levels_labels function implemented within the pixiedust package (https://github.com/nutterb/pixiedust/blob/master/R/tidy_levels_labels.R).

Value

The data.frame produced by tidy with these additional columns:

variable

variable name

variable_label

label of the variable (if available), variable name otherwise

level

for factors, the corresponding level

level_detail

level with indication of the reference level, eg. B vs. A

label

equal to level_detail if it's a factor, to variable_label otherwise

Note

This function requires the broom and dplyr packages.

See also

Examples

reg <- lm(Sepal.Length ~ Sepal.Width + Petal.Length * Species, data = iris)
tidy_detailed(reg)

#> Error: `tidy_detailed()` was deprecated in JLutils 1.21.0 and is now defunct.
#> Please use `broom.helpers::tidy_plus_plus()` instead.

d <- as.data.frame(Titanic)
reg2 <- glm(Survived ~ Sex + Age + Class, family = binomial, data = d, weights = d$Freq)
tidy_detailed(reg2, conf.int = TRUE, exponentiate = TRUE)

#> Error: `tidy_detailed()` was deprecated in JLutils 1.21.0 and is now defunct.
#> Please use `broom.helpers::tidy_plus_plus()` instead.