Tidy data with variable names and factor levels

This function extends the output of tidy function from broom package by providing additional column indicating variables and factors names.

tidy_detailed(x, ...)

Arguments

x
model to tidy

...
extra arguments passed to tidy function from broom package

Source

This function has been adapted from Benjamin Nutter's tidy_levels_labels function implemented within the pixiedust package (https://github.com/nutterb/pixiedust/blob/master/R/tidy_levels_labels.R).

Value

The data.frame produced by tidy with these additional columns:

variable

variable name

variable_label

label of the variable (if available), variable name otherwise

level

for factors, the corresponding level

level_detail

level with indication of the reference level, eg. B vs. A

label

equal to level_detail if it's a factor, to variable_label otherwise

Note

This function requires the broom and dplyr packages.

See also

Examples

reg <- lm(Sepal.Length ~ Sepal.Width + Petal.Length * Species, data = iris)
tidy_detailed(reg)

#> Le chargement a nécessité le package : broom

#>                             term   estimate  std.error statistic      p.value
#> 1                    (Intercept)  2.9257321 0.43637647  6.704606 4.338251e-10
#> 2                   Petal.Length  0.3677469 0.25359434  1.450138 1.492097e-01
#> 3 Petal.Length:Speciesversicolor  0.2920936 0.26829782  1.088692 2.781212e-01
#> 4  Petal.Length:Speciesvirginica  0.5225336 0.26411484  1.978433 4.980125e-02
#> 5                    Sepal.Width  0.4500064 0.08060809  5.582646 1.153465e-07
#> 6              Speciesversicolor -1.0471703 0.56084638 -1.867125 6.393137e-02
#> 7               Speciesvirginica -2.6188882 0.58445261 -4.480925 1.510966e-05
#>               variable       variable_label                   level
#> 1                 <NA>                 <NA>                    <NA>
#> 2         Petal.Length         Petal.Length                        
#> 3 Petal.Length:Species Petal.Length:Species Petal.Length:versicolor
#> 4 Petal.Length:Species Petal.Length:Species  Petal.Length:virginica
#> 5          Sepal.Width          Sepal.Width                        
#> 6              Species              Species              versicolor
#> 7              Species              Species               virginica
#>                         level_detail                              label
#> 1                               <NA>                               <NA>
#> 2                                                          Petal.Length
#> 3 Petal.Length:versicolor vs. setosa Petal.Length:versicolor vs. setosa
#> 4  Petal.Length:virginica vs. setosa  Petal.Length:virginica vs. setosa
#> 5                                                           Sepal.Width
#> 6              versicolor vs. setosa              versicolor vs. setosa
#> 7               virginica vs. setosa               virginica vs. setosa

d <- as.data.frame(Titanic)
reg2 <- glm(Survived ~ Sex + Age + Class, family = binomial, data = d, weights = d$Freq)
tidy_detailed(reg2, conf.int = TRUE, exponentiate = TRUE)

#>          term   estimate std.error  statistic      p.value  conf.low  conf.high
#> 1 (Intercept)  1.9844057 0.2729942   2.510381 1.206011e-02 1.1611867  3.3899109
#> 2    AgeAdult  0.3459219 0.2440256  -4.350127 1.360589e-05 0.2140861  0.5578928
#> 3    Class2nd  0.3612825 0.1959975  -5.194428 2.053497e-07 0.2453064  0.5291859
#> 4    Class3rd  0.1690159 0.1715665 -10.361940 3.693921e-25 0.1203213  0.2358425
#> 5   ClassCrew  0.4241466 0.1573389  -5.451140 5.004800e-08 0.3115232  0.5774746
#> 6   SexFemale 11.2465380 0.1404101  17.235662 1.434015e-66 8.5727281 14.8700792
#>   variable variable_label  level    level_detail           label
#> 1     <NA>           <NA>   <NA>            <NA>            <NA>
#> 2      Age            Age  Adult Adult vs. Child Adult vs. Child
#> 3    Class          Class    2nd     2nd vs. 1st     2nd vs. 1st
#> 4    Class          Class    3rd     3rd vs. 1st     3rd vs. 1st
#> 5    Class          Class   Crew    Crew vs. 1st    Crew vs. 1st
#> 6      Sex            Sex Female Female vs. Male Female vs. Male